Moved some components into a subcomponent folder
