+++ src/components/subcomponents/perkinfo.vue
